requirements.txt
numpy>=1.20.0
matplotlib>=3.3.0
scikit-image>=0.18.0
scikit-learn>=1.0.0
scipy>=1.7.0,<1.11.0
progressbar2>=4.0.0
Pillow>=8.0.0
pandas>=1.3.0
tensorflow>=2.10.0
scikit-plot>=0.3.7
seaborn>=0.11.0


